Tutorial 11: extending the registration form
Added some fields to registration.

This REALLY needs some validating though... if you enter a first_name that is >30 characters, it silently fails
(which is better than actually succeeding, I suppose).  It also allows duplicate email etc.

But great proof of concept.  I am worried because 1.8 is said to have really improved on the auth so
we may find that this entire portion is outdated.