Tutorial 14: File Uploading and Updating database schema
Same version problems as we had with STATICFILES_DIRS.  It's "WORKING", but I am not certain that all is perfect.
I suspect that with a bit more experience we'll want to revisit this, but for now it gets the job done.

We do not need to "dump" our database, just migrate normally and all old data will be retained.

When creating the new "thumbnail" field in the Article model, you must give it a default value or it fails.